move some stuff to new gui2.common
